Add AOT-friendly JSON context and split test files
- Add TestResultEventJsonContext (JsonSerializerContext) for AOT compatibility
- Add ToString() override that serializes via the AOT context
- Update Parse() to use the AOT context for deserialization
- Simplify EventStreamFormatter to use ToString() instead of JsonSerializer
- Split EventStreamTests.cs into separate files per test class:
  - EventStreamFormatterTests.cs
  - TestResultEventParseTests.cs (with new ToString round-trip tests)
  - TestResultEventToInfoTests.cs
  - TestResultEventFactoryTests.cs
  - TestResultEventRoundTripTests.cs
  - TestHelpers.cs (shared factory)
- All 170 tests pass (135 VisualRunners + 35 CLI)
